get ego X cell reports

Produces a dataframe that has a row for each respondent X cell containing respondent's reported deaths and exposure among siblings in the cell.

get_ec_reports(
  esc.dat,
  ego.id,
  sib.dat,
  sib.frame.indicator,
  cell.vars,
  weights,
  discretize.exp = FALSE
)

Arguments

esc.dat
Dataset with a row for each respondent X sibling X cell, likely produced by get_esc_reports

ego.id
String with the name of the column in esc.dat containing the survey respondent's id

sib.dat
Dataset with a row for each reported sibling, likely produced by TODO

sib.frame.indicator
String with the name of the column in sib.dat containing a 0/1 coded variable indicating whether or not each sib is in the frame population

cell.vars
a vector with Strings containing the names of columns in esc.dat identifying the cells to group reports by (typically age group, sex, time period)

weights
String with the name of the column in esc.dat that has the survey weights

discretize.exp
Boolean for whether or not expsoure should be discretized. Not yet implemented.

Value

A dataframe that has a row for each respondent X cell containing respondent's reported deaths and exposure among siblings in the cell.

Examples

  # TODO - add example