prepare a DHS dataset for analysis

prepare a DHS dataset for analysis

prep_dhs_sib_histories(
  df,
  varmap = sibhist_varmap_dhs6,
  keep_missing = FALSE,
  keep_varmap_only = FALSE,
  verbose = TRUE
)

Arguments

df
the raw DHS dataset (indvidual recode)

varmap
see Details

keep_missing
should we keep reported sibs that are missing sex or survival status?

keep_varmap_only
should we only keep ego variables mentioned in the varmap? (Default: FALSE)

verbose
report detailed summaries?

Value

a list; see Details

Details

varmap should be a dataframe with columns TODO. For respondents, you should be sure to include

survey (the survey id, usually a country code plus one digit)

caseid (the respondent id)

wwgt (the sampling weight for women)

psu (the primary sampling unit)

doi (the date of the interview)

For siblings, you should be sure to include

sib.death.date (the date of the sibling's death)

sib.alive (whether or not the sib is alive)

sib.sex (the sex of the sibling).

The default varmap is sibhist_varmap_dhs6, which is included with the package.

Returns a list whose entries include

ego.dat - dataset with information about the survey respondents

sib.dat - dataset with information about the reported siblings

summ - a one-row tibble with a summary of the data

Examples

  # TODO - write example code